requirements.txt
streamlit
ultralytics
opencv-python
tensorflow
matplotlib
Pillow